requirements.txt
PySide6
qasync
google-generativeai
python-dotenv
Pygments
Markdown
openai
ollama
qtawesome